+++ backend/src/main/kotlin/backend/bonuses/AwardService.kt
@@ -1,0 +1,4 @@
+package backend.bonuses
+
+class AwardService {
+}
