+++ backend/src/main/kotlin/backend/bonuses/BonusesConfiguration.kt
@@ -1,0 +1,4 @@
+package backend.bonuses
+
+class BonusesConfiguration {
+}
