+++ backend/src/main/kotlin/backend/bonuses/BonusesRepository.kt
@@ -1,0 +1,9 @@
+package backend.bonuses
+
+import backend.bonuses.Bonuses
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface BonusesRepository : JpaRepository<Bonuses, Long>{
+}
