+++ backend/src/main/kotlin/backend/categories/CategoryRepository.kt
@@ -1,0 +1,9 @@
+package backend.categories
+
+import backend.categories.Categories
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface CategoryRepository: JpaRepository<Categories, Long> {
+}
