+++ backend/src/main/kotlin/backend/categories/CategoryService.kt
@@ -1,0 +1,15 @@
+package backend.categories
+import jakarta.transaction.Transactional
+import org.springframework.stereotype.Service
+
+@Service
+class CategoryService(
+    private val categoryRepository: CategoryRepository
+) {
+
+    @Transactional
+    fun createCategory(categoryName: String): Categories {
+        val category = Categories(categoryName = categoryName)
+        return categoryRepository.save(category)
+    }
+}
