+++ backend/src/main/kotlin/backend/chestHistory/ChestHistoryRepository.kt
@@ -1,0 +1,9 @@
+package backend.chestHistory
+
+import backend.chestHistory.ChestHistory
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface ChestHistoryRepository : JpaRepository<ChestHistory, Long> {
+}
