+++ backend/src/main/kotlin/backend/points/PointsRepository.kt
@@ -1,0 +1,9 @@
+package backend.points
+
+import backend.points.Points
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface PointsRepository : JpaRepository<Points, Long> {
+}
