--- backend/src/main/kotlin/backend/repository/BonusesRepository.kt
@@ -1,9 +1,0 @@
-package backend.repository
-
-import backend.entity.Bonuses
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface BonusesRepository : JpaRepository<Bonuses, Long>{
-}
