--- backend/src/main/kotlin/backend/repository/ChestAwardRepository.kt
@@ -1,9 +1,0 @@
-package backend.repository
-
-import backend.entity.ChestAward
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface ChestAwardRepository : JpaRepository<ChestAward, Long> {
-}
