--- backend/src/main/kotlin/backend/repository/ChestHistoryRepository.kt
@@ -1,9 +1,0 @@
-package backend.repository
-
-import backend.entity.ChestHistory
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface ChestHistoryRepository : JpaRepository<ChestHistory, Long> {
-}
