--- backend/src/main/kotlin/backend/repository/ChestsRepository.kt
@@ -1,9 +1,0 @@
-package backend.repository
-
-import backend.entity.Chests
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface ChestsRepository : JpaRepository<Chests, Long> {
-}
