--- backend/src/main/kotlin/backend/repository/PointsRepository.kt
@@ -1,9 +1,0 @@
-package backend.repository
-
-import backend.entity.Points
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface PointsRepository : JpaRepository<Points, Long> {
-}
