--- backend/src/main/kotlin/backend/repository/SubcategoriesRepository.kt
@@ -1,9 +1,0 @@
-package backend.repository
-
-import backend.entity.Subcategories
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface SubcategoriesRepository : JpaRepository<Subcategories, Long> {
-}
