--- backend/src/main/kotlin/backend/repository/UsersRepository.kt
@@ -1,9 +1,0 @@
-package backend.repository
-
-import backend.entity.Users
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface UsersRepository : JpaRepository<Users, Long> {
-}
