--- backend/src/main/kotlin/backend/service/PointsService.kt
@@ -1,48 +1,0 @@
-package backend.service
-
-import backend.entity.Points
-import backend.entity.Users
-import backend.entity.Subcategories
-import jakarta.persistence.EntityManager
-import jakarta.transaction.Transactional
-import org.springframework.stereotype.Service
-
-@Service
-class PointsService(
-    private val entityManager: EntityManager
-) {
-    @Transactional
-    fun createPoints(userId: Users, fromWho: String, howMany: Long, subcategory: Subcategories): Points {
-        val points = Points(userId = userId, fromWho = fromWho, howMany = howMany, subcategory = subcategory)
-        entityManager.persist(points)
-        return points
-    }
-
-    @Transactional
-    fun updatePoints(pointsId: Long, userId: Users, fromWho: String, howMany: Long, subcategory: Subcategories): Points? {
-        val points = entityManager.find(Points::class.java, pointsId)
-        if (points != null) {
-            points.userId = userId
-            points.fromWho = fromWho
-            points.howMany = howMany
-            points.subcategory = subcategory
-            entityManager.merge(points)
-        }
-        return points
-    }
-
-    @Transactional
-    fun findPointsById(pointsId: Long): Points? {
-        return entityManager.find(Points::class.java, pointsId)
-    }
-
-    @Transactional
-    fun deletePoints(pointsId: Long): Boolean {
-        val points = entityManager.find(Points::class.java, pointsId)
-        if (points != null) {
-            entityManager.remove(points)
-            return true
-        }
-        return false
-    }
-}
