+++ backend/src/main/kotlin/backend/users/Users.kt
@@ -1,0 +1,23 @@
+package backend.users
+import jakarta.persistence.*
+
+
+@Entity
+@Table(name = "users")
+class Users (
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "userId")
+    val userId: Long = 0,
+
+    @Column(name = "nick", nullable = false)
+    var nick: String,
+
+    @Column(name = "role", nullable = false)
+    var role: String,
+){
+    constructor() : this(
+        nick = "",
+        role = ""
+    )
+}
