+++ backend/src/main/kotlin/backend/users/UsersRepository.kt
@@ -1,0 +1,9 @@
+package backend.users
+
+import backend.users.Users
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface UsersRepository : JpaRepository<Users, Long> {
+}
